pipeline/requirements.txt
boto3
openpyxl
pandas
redshift_connector
python-dotenv
xlrd
altair
streamlit

pyarrow
fastparquet
